requirements.txt
opencv-python
opencv-contrib-python
langchain
langchain-openai
langchain_google_genai
langchain_groq
langchain-community
python-dotenv
sounddevice
pyaudio
piper-tts
soundfile
SpeechRecognition
SpeechRecognition[whisper-local]
fastapi
uvicorn
fastapi[all]
ffmpeg-python